threaded loading code and associated/extensive tweaks. unified image loading code a little between renderers. support switching worldmodel in csqc. also associated bugfixes.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4758 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/qclib/qcd.h
@@ -2,6 +2,6 @@ pbool QC_decodeMethodSupported(int method);
 char *QC_decode(progfuncs_t *progfuncs, int complen, int len, int method, char *info, char *buffer);
 int QC_encode(progfuncs_t *progfuncs, int len, int method, char *in, int handle);
 
-char *PDECL filefromprogs(pubprogfuncs_t *progfuncs, progsnum_t prnum, char *fname, int *size, char *buffer);
-char *filefromnewprogs(pubprogfuncs_t *progfuncs, char *prname, char *fname, int *size, char *buffer);//fixme - remove parm 1
+char *PDECL filefromprogs(pubprogfuncs_t *progfuncs, progsnum_t prnum, char *fname, size_t *size, char *buffer);
+char *filefromnewprogs(pubprogfuncs_t *progfuncs, char *prname, char *fname, size_t *size, char *buffer);//fixme - remove parm 1
